refactor: 패키지 구조 수정 & 클래스명 변경 (#230)
