+++ backend/src/main/java/com/celuveat/common/exception/ExceptionControllerAdvice.java
@@ -23,7 +23,7 @@ public class ExceptionControllerAdvice {
     @ExceptionHandler(BaseException.class)
     ResponseEntity<ExceptionResponse> handleException(BaseException e) {
         BaseExceptionType type = e.exceptionType();
-        LogContext logContext = logContextHolder.get();
+        LogContext logContext = logContextHolder.getOrCreate();
         log.info("[BAD_REQUEST][{}] 잘못된 요청이 들어왔습니다. 내용:  {}", logContext.logId(), type.errorMessage());
         return ResponseEntity.status(type.httpStatus())
                 .body(new ExceptionResponse(type.errorMessage()));
@@ -34,7 +34,7 @@ public class ExceptionControllerAdvice {
         String errorMessage = e.getFieldErrors().stream()
                 .map(it -> it.getField() + " : " + it.getDefaultMessage())
                 .collect(Collectors.joining("  "));
-        LogContext logContext = logContextHolder.get();
+        LogContext logContext = logContextHolder.getOrCreate();
         log.info("[BAD_REQUEST][{}] 잘못된 요청이 들어왔습니다. 내용:  {}", logContext.logId(), errorMessage);
         return ResponseEntity.status(BAD_REQUEST)
                 .body(new ExceptionResponse(errorMessage));
@@ -42,7 +42,7 @@ public class ExceptionControllerAdvice {
 
     @ExceptionHandler(MissingServletRequestParameterException.class)
     ResponseEntity<ExceptionResponse> handleMissingParams(MissingServletRequestParameterException e) {
-        LogContext logContext = logContextHolder.get();
+        LogContext logContext = logContextHolder.getOrCreate();
         String errorMessage = e.getParameterName() + " 값이 누락 되었습니다.";
         log.info("[BAD_REQUEST][{}] 잘못된 요청이 들어왔습니다. 내용:  {}", logContext.logId(), errorMessage);
         return ResponseEntity.status(BAD_REQUEST)
@@ -51,7 +51,7 @@ public class ExceptionControllerAdvice {
 
     @ExceptionHandler(Exception.class)
     ResponseEntity<ExceptionResponse> handleException(Exception e) {
-        LogContext logContext = logContextHolder.get();
+        LogContext logContext = logContextHolder.getOrCreate();
         log.error("[ERROR][{}] 예상하지 못한 예외가 발생했습니다.", logContext.logId(), e);
         return ResponseEntity.internalServerError()
                 .body(new ExceptionResponse("알 수 없는 오류가 발생했습니다."));
