+++ backend/src/main/java/com/celuveat/common/log/request/logid/AuthenticatedLogId.java
@@ -1,4 +1,4 @@
-package com.celuveat.common.log.request;
+package com.celuveat.common.log.request.logid;
 
 public class AuthenticatedLogId extends RequestLogId {
