+++ backend/src/main/java/com/celuveat/common/log/request/logid/RequestLogIdHolder.java
@@ -1,4 +1,4 @@
-package com.celuveat.common.log.request;
+package com.celuveat.common.log.request.logid;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
