+++ backend/src/main/java/com/celuveat/common/log/request/messagebody/MessageBodyReader.java
@@ -1,4 +1,4 @@
-package com.celuveat.common.log.request;
+package com.celuveat.common.log.request.messagebody;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -7,7 +7,7 @@ import java.io.IOException;
 import java.util.stream.Collectors;
 import org.springframework.web.util.ContentCachingResponseWrapper;
 
-public class RequestBodyReader {
+public class MessageBodyReader {
 
     @NotNull
     public static String readBody(HttpServletRequest request) {
